First Stable Build
All bugs sorted for the time being.

The only thing that could cause an issue would be very slow loading of a website.

THIS SCRIPT HAS A MINIMUM OF 10Mbps Up/Down TO WORK BUT SHOULD IDEALLY HAVE 100Mbps OF BANDWIDTH AVAILABLE. (Aka: Use ethernet)

THIS SCRIPT NOW REQUIRES A HEADLESS SETUP (Meaning anyone running this on a Libre should grab an HDMI Emulation port filler) - MINIMUM RESOLUTION IS 1280x720!